naturalql/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
groq==0.9.0
sqlparse==0.5.0
psycopg2-binary==2.9.9
mysql-connector-python==8.4.0
python-multipart==0.0.9
